// src/app/models/call.ts
export interface Call{
    
    id?:                   any;
    openDate?:          string;
    closeDate?:         string;
    priority:           String;
    status:             String;
    title:              string;
    description:        string;
    technician:            any;
    client:                any;
    clientName:         string;
    technicianName:     string;
    
}